+++ src/Bem/Utl/Utl.hs
@@ -1,21 +1,33 @@
 {-# LANGUAGE EmptyDataDeriving #-}
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE KindSignatures #-}
+{-# LANGUAGE RecordWildCards #-}
 
 
 -- | high-level extra utilities
 module Bem.Utl.Utl where
 
 
-import Bem.Cls.Gen.Gen
+import qualified Bem.Cls.Gen.Gen as Gen
 import Bem.Utl.Intr
-
-import Control.Monad.Reader
-import Data.Kind
-
 import Bem.Cfg.Cfg
+import Control.Monad.Reader
+import qualified Bem.Cls.Gen.Cfg as Cfg
 import qualified Bem.Cls.Gen.Intr as IntrGen
 
+import Data.Kind
+
+
+-- | the configurable extra class generators
+data Gens = Gens { _genBlk :: FromBlkElem Class
+                 , _genElem :: forall b . FromFullElem b Class
+                 , _genNoBlkModsBlk
+                       :: forall b
+                       . FromBlk b (FromFullElem b Class)
+                 , _genNoElemModsBlk :: FromBlkNoModsElem Class
+                 , _genNoModsBlk :: FromNoModsBlkNoModsElem Class
+                 , _genNoModsElem :: forall b . FromElem b Class
+                 }
 
 -- | Denote absent elements.
 data NoElem m
@@ -28,21 +40,21 @@ data NoMod deriving Show
 genNoBlkModsBlk :: FromNoModsBlkElem Class
 genNoBlkModsBlk blk prntBlk elem' elemMods
     =
-    genBlkElem blk [] prntBlk elem' elemMods
+    Gen.genBlkElem blk [] prntBlk elem' elemMods
 
 -- | Generate a class of a block and element that is without its modifiers.
 genNoElemModsBlk :: FromBlkNoModsElem Class
 genNoElemModsBlk blk blkMods prntBlk elem'
     =
-    genBlkElem blk blkMods prntBlk elem' []
+    Gen.genBlkElem blk blkMods prntBlk elem' []
 
 -- | Generate a class of a block and element without any modifiers.
 genNoModsBlk :: FromNoModsBlkNoModsElem Class
-genNoModsBlk blk prntBlk elem' = genBlkElem blk [] prntBlk elem' []
+genNoModsBlk blk prntBlk elem' = Gen.genBlkElem blk [] prntBlk elem' []
 
 -- | Generate a class of an element that is without its modifiers.
 genNoModsElem :: FromElem b Class
-genNoModsElem prntBlk elem' = genElem prntBlk elem' []
+genNoModsElem prntBlk elem' = Gen.genElem prntBlk elem' []
 
 -- | Decorate a single block.
 decorSingleton :: (Show (b e m)) => b (e :: Type -> Type) m -> Class
@@ -53,3 +65,33 @@ decorSingleton = (`runReader` defCfg) . (IntrGen.decor . IntrGen.Blk)
                  , _partSep = "-"
                  , _partsAreCptled = False
                  }
+
+-- | Initialise the utility configurable class generators with a configuration.
+init :: Cfg -> Gens
+init cfg
+    =
+    Gens { _genBlk = _genBlk
+         , _genElem = _genElem
+         , _genNoBlkModsBlk
+               =
+               \blk prntBlk elem' elemMods
+               ->
+               runReader (IntrGen.genBlk blk [] prntBlk elem' elemMods) cfg
+         , _genNoElemModsBlk
+               =
+               \blk blkMods prntBlk elem'
+               ->
+               runReader (IntrGen.genBlk blk blkMods prntBlk elem' []) cfg
+         , _genNoModsBlk
+               =
+               \blk prntBlk elem'
+               ->
+               runReader (IntrGen.genBlk blk [] prntBlk elem' []) cfg
+         , _genNoModsElem
+               =
+               \prntBlk elem'
+               ->
+               runReader (IntrGen.genElem prntBlk elem' []) cfg
+         }
+  where
+    Cfg.Gens {..} = Cfg.init cfg